moved types file to the main folder
