+++ src/actions/authedUser.ts
@@ -1,4 +1,4 @@
-import { AuthedUsersActionTypes } from './types';
+import { AuthedUsersActionTypes } from '../types';
 export const SET_AUTHED_USER = 'SET_AUTHED_USER';
 export const LOGOUT_AUTHED_USER = 'LOGOUT_AUTHED_USER';
